moved blogTable into pages
